# Optimized nextNonEquivalentValue(), and consequently ValueAtPercentile, ValueAtPercentiles...