--- a/example_hdr_test.go
+++ b/example_hdr_test.go
@@ -88,28 +88,27 @@ func ExampleHistogram_PercentilesPrint() {
 
 // When doing an percentile analysis we normally require more than one percentile to be calculated for the given histogram.
 //
-// When that is the case ValueAtQuantiles() will deeply optimize the total time to retrieve the percentiles vs the other option
+// When that is the case ValueAtPercentiles() will deeply optimize the total time to retrieve the percentiles vs the other option
 // which is multiple calls to ValueAtQuantile().
 //
-// NOTE: The ValueAtQuantiles optimization implies that you pass an ORDERED (smaller to greater) slice of percentiles
 // nolint
-func ExampleHistogram_ValueAtQuantiles() {
+func ExampleHistogram_ValueAtPercentiles() {
 	histogram := hdrhistogram.New(1, 30000000, 3)
 
 	for i := 0; i < 1000000; i++ {
 		histogram.RecordValue(int64(i))
 	}
 
-	percentileValues := histogram.ValueAtQuantiles([]float64{50.0, 95.0, 99.0, 99.0})
-	fmt.Printf("Percentile 50: %d\n", percentileValues[0])
-	fmt.Printf("Percentile 95: %d\n", percentileValues[1])
-	fmt.Printf("Percentile 99: %d\n", percentileValues[2])
-	fmt.Printf("Percentile 99.9: %d\n", percentileValues[3])
+	percentileValuesMap := histogram.ValueAtPercentiles([]float64{50.0, 95.0, 99.0, 99.9})
+	fmt.Printf("Percentile 50: %d\n", percentileValuesMap[50.0])
+	fmt.Printf("Percentile 95: %d\n", percentileValuesMap[95.0])
+	fmt.Printf("Percentile 99: %d\n", percentileValuesMap[99.0])
+	fmt.Printf("Percentile 99.9: %d\n", percentileValuesMap[99.9])
 
 	// Output:
 	// Percentile 50: 500223
 	// Percentile 95: 950271
 	// Percentile 99: 990207
-	// Percentile 99.9: 990719
+	// Percentile 99.9: 999423
 
 }