--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/HdrHistogram/hdrhistogram-go
 
-go 1.14
+go 1.11
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
@@ -8,7 +8,6 @@ require (
 	github.com/kr/text v0.2.0 // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/exp v0.0.0-20210126221216-84987778548c // indirect
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 	gonum.org/v1/gonum v0.8.2
 	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect