--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/HdrHistogram/hdrhistogram-go
 
-go 1.11
+go 1.14
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect