I've made the changes on top of PR #44 ( but the optimizations are independent, I just reused the tests )... We should review #44 prior this one...

This PR focus on two internal API changes that should boost the performance of functions that rely on nextNonEquivalentValue() significantly without any change on the exposed methods, and even a code reduction...

### Optimization 1
Calculate only once the bucket index within nextNonEquivalentValue, given `getBucketIndex()` is the method consuming most of the CPU, as seen bellow:

![image](https://user-images.githubusercontent.com/5832149/107864099-37a68b80-6e51-11eb-8afb-f3025120e8b2.png)

This allowed to move 101 microseconds per call of ValueAtPercentiles, to 67 microseconds, as seen bellow:

```
fco@fcos-Air hdrhistogram-go % go test -bench=BenchmarkHistogramValueAtPercentilesGivenPercentileSlice -benchmem -benchtime=60s -benchmem -cpuprofile=cpu.opt2.out 
goos: darwin
goarch: arm64
pkg: github.com/HdrHistogram/hdrhistogram-go
BenchmarkHistogramValueAtPercentilesGivenPercentileSlice-8        699392            101208 ns/op             248 B/op          4 allocs/op
BenchmarkHistogramValueAtPercentilesGivenPercentileSlice-8        984460             67611 ns/op             248 B/op          4 allocs/op
```


------------------------


### Optimization 2, use Bits package LeadingZeros64() instead of our own implementation

After the above optimization we see that bitLen is showing as one of the hot methods and within the critical code path.
![image](https://user-images.githubusercontent.com/5832149/107864271-79840180-6e52-11eb-8f3c-fbf9ae46c17e.png)

Given that calculating the number of leading 0-bits is a very common problem I believe we should reuse a battle tested implemntation that not only with reduce a bit the total CPU time of methods that call nextNonEquivalentValue but also reduce the code that the package implements himself. 

This allowed to move 67 microseconds per call of ValueAtPercentiles, to 65 microseconds, as seen bellow:

```
fco@fcos-Air hdrhistogram-go % go test -bench=BenchmarkHistogramValueAtPercentilesGivenPercentileSlice -benchmem -benchtime=60s -benchmem -cpuprofile=cpu.opt2.out 
goos: darwin
goarch: arm64
pkg: github.com/HdrHistogram/hdrhistogram-go
BenchmarkHistogramValueAtPercentilesGivenPercentileSlice-8        984460             67611 ns/op             248 B/op          4 allocs/op
BenchmarkHistogramValueAtPercentilesGivenPercentileSlice-8       1000000             65452 ns/op             248 B/op          4 allocs/op


```